html/semantics/embedded-content-0/media-elements/preload-metadata-buffered-webm.html

<!doctype html>
<!-- This file was generated from 'preload-metadata-buffered.tpl' using 'tpl2tc.py'. Please do not edit manually! -->
<title>HTMLMediaElement.preload:metadata 'buffered' check - webm</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/common.js"></script>
<p id="log">FAIL (script didn't run)</p>
<audio preload="metadata" controls></audio>
<video preload="metadata" controls></video>
<script>
var tests = init_tests("HTMLAudioElement.preload:metadata 'buffered' check - webm", "HTMLVideoElement.preload:metadata 'buffered' check - webm", {timeout:10000});
tests.forEach(function(vars) {
    var t = vars[0];
    t.step(function() {
	var tag_name = vars[1];
	var events_actual = [];
	var node = document.getElementsByTagName(tag_name)[0];
	startTest();

	function startTest() {
	    var events  = ['loadstart', 'progress', 'suspend', 'abort', 'error', 'loadedmetadata', 'loadeddata', 'canplay', 'canplaythrough'];
	    for (var e in events) {
		node.addEventListener(events[e], handleEvent, false);
	    }
	    node.src = '../resources/preload.webm?pipe=trickle(100000:d1:r2)&nocache=' + Math.random();
	}

	function handleEvent(e) {
	    if (e.type == 'error' && !node.canPlayType(tag_name + '/webm')) {
		t.step(function() {
		    assert_unreached("Cannot play '" + tag_name + "/webm'");
		    t.done();
		});
	    }
	    if (e.type == 'suspend') endTest();
	}

	function endTest() {
	    t.step(function() {
		assert_equals(node.buffered.length, 1, "'buffered' should have a length of 1");
		assert_equals(node.buffered.start(0), 0, "The start of buffered timerange should be 0");
		assert_true(node.buffered.end(0) < 60, "The end of buffered timerange should be less than 60 seconds" );
		t.done();
	    });
	}
    });
});
</script>
